Fix steemconnect scope
I'm made a fix on the steemconnect scope cuz the latest version of SC is more strict about it and may break the login. The scope "login" should only be used if none of the posting operation is used, and the "offline" scope is only used if you want to generate refresh token, which is not the case as far as i understand.